[rust-axum] Fix: add samples output
