+++ samples/server/petstore/rust-axum/output/multipart-v3/src/apis/mod.rs
@@ -1,0 +1,1 @@
+pub mod default;
