+++ samples/server/petstore/rust-axum/output/openapi-v3/.openapi-generator/FILES
@@ -1,6 +1,9 @@
 .gitignore
 Cargo.toml
 README.md
+src/apis/default.rs
+src/apis/mod.rs
+src/apis/repo.rs
 src/header.rs
 src/lib.rs
 src/models.rs
