+++ samples/server/petstore/rust-axum/output/openapi-v3/src/apis/mod.rs
@@ -1,0 +1,2 @@
+pub mod default;
+pub mod repo;
