+++ samples/server/petstore/rust-axum/output/openapi-v3/src/server/mod.rs
@@ -10,24 +10,13 @@ use validator::{Validate, ValidationErrors};
 use crate::{header, types::*};
 
 #[allow(unused_imports)]
-use crate::models;
-
-use crate::{
-    AnyOfGetResponse, Api, CallbackWithHeaderPostResponse, ComplexQueryParamGetResponse,
-    CreateRepoResponse, EnumInPathPathParamGetResponse, GetRepoInfoResponse,
-    JsonComplexQueryParamGetResponse, MandatoryRequestHeaderGetResponse, MergePatchJsonGetResponse,
-    MultigetGetResponse, MultipleAuthSchemeGetResponse, OneOfGetResponse,
-    OverrideServerGetResponse, ParamgetGetResponse, ReadonlyAuthSchemeGetResponse,
-    RegisterCallbackPostResponse, RequiredOctetStreamPutResponse, ResponsesWithHeadersGetResponse,
-    Rfc7807GetResponse, UntypedPropertyGetResponse, UuidGetResponse, XmlExtraPostResponse,
-    XmlOtherPostResponse, XmlOtherPutResponse, XmlPostResponse, XmlPutResponse,
-};
+use crate::{apis, models};
 
 /// Setup API Server.
 pub fn new<I, A>(api_impl: I) -> Router
 where
     I: AsRef<A> + Clone + Send + Sync + 'static,
-    A: Api + 'static,
+    A: apis::default::Default + apis::repo::Repo + 'static,
 {
     // build our application with a route
     Router::new()
@@ -107,7 +96,7 @@ async fn any_of_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = any_of_get_validation(query_params);
 
@@ -127,7 +116,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            AnyOfGetResponse::Status200_Success(body) => {
+            apis::default::AnyOfGetResponse::Status200_Success(body) => {
                 let mut response = response.status(200);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -150,7 +139,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            AnyOfGetResponse::Status201_AlternateSuccess(body) => {
+            apis::default::AnyOfGetResponse::Status201_AlternateSuccess(body) => {
                 let mut response = response.status(201);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -173,7 +162,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            AnyOfGetResponse::Status202_AnyOfSuccess(body) => {
+            apis::default::AnyOfGetResponse::Status202_AnyOfSuccess(body) => {
                 let mut response = response.status(202);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -229,7 +218,7 @@ async fn callback_with_header_post<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = callback_with_header_post_validation(query_params);
 
@@ -249,7 +238,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            CallbackWithHeaderPostResponse::Status204_OK => {
+            apis::default::CallbackWithHeaderPostResponse::Status204_OK => {
                 let mut response = response.status(204);
                 response.body(Body::empty())
             }
@@ -286,7 +275,7 @@ async fn complex_query_param_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = complex_query_param_get_validation(query_params);
 
@@ -306,7 +295,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            ComplexQueryParamGetResponse::Status200_Success => {
+            apis::default::ComplexQueryParamGetResponse::Status200_Success => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -343,7 +332,7 @@ async fn enum_in_path_path_param_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = enum_in_path_path_param_get_validation(path_params);
 
@@ -363,7 +352,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            EnumInPathPathParamGetResponse::Status200_Success => {
+            apis::default::EnumInPathPathParamGetResponse::Status200_Success => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -400,7 +389,7 @@ async fn json_complex_query_param_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = json_complex_query_param_get_validation(query_params);
 
@@ -420,7 +409,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            JsonComplexQueryParamGetResponse::Status200_Success => {
+            apis::default::JsonComplexQueryParamGetResponse::Status200_Success => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -457,7 +446,7 @@ async fn mandatory_request_header_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     // Header parameters
     let header_params = {
@@ -510,7 +499,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            MandatoryRequestHeaderGetResponse::Status200_Success => {
+            apis::default::MandatoryRequestHeaderGetResponse::Status200_Success => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -542,7 +531,7 @@ async fn merge_patch_json_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = merge_patch_json_get_validation();
 
@@ -562,7 +551,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            MergePatchJsonGetResponse::Status200_Merge(body) => {
+            apis::default::MergePatchJsonGetResponse::Status200_Merge(body) => {
                 let mut response = response.status(200);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -613,7 +602,7 @@ async fn multiget_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = multiget_get_validation();
 
@@ -630,7 +619,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            MultigetGetResponse::Status200_JSONRsp(body) => {
+            apis::default::MultigetGetResponse::Status200_JSONRsp(body) => {
                 let mut response = response.status(200);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -653,7 +642,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            MultigetGetResponse::Status201_XMLRsp(body) => {
+            apis::default::MultigetGetResponse::Status201_XMLRsp(body) => {
                 let mut response = response.status(201);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -669,7 +658,7 @@ where
                 let body_content = body;
                 response.body(Body::from(body_content))
             }
-            MultigetGetResponse::Status202_OctetRsp(body) => {
+            apis::default::MultigetGetResponse::Status202_OctetRsp(body) => {
                 let mut response = response.status(202);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -685,7 +674,7 @@ where
                 let body_content = body.0;
                 response.body(Body::from(body_content))
             }
-            MultigetGetResponse::Status203_StringRsp(body) => {
+            apis::default::MultigetGetResponse::Status203_StringRsp(body) => {
                 let mut response = response.status(203);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -701,7 +690,7 @@ where
                 let body_content = body;
                 response.body(Body::from(body_content))
             }
-            MultigetGetResponse::Status204_DuplicateResponseLongText(body) => {
+            apis::default::MultigetGetResponse::Status204_DuplicateResponseLongText(body) => {
                 let mut response = response.status(204);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -724,7 +713,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            MultigetGetResponse::Status205_DuplicateResponseLongText(body) => {
+            apis::default::MultigetGetResponse::Status205_DuplicateResponseLongText(body) => {
                 let mut response = response.status(205);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -747,7 +736,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            MultigetGetResponse::Status206_DuplicateResponseLongText(body) => {
+            apis::default::MultigetGetResponse::Status206_DuplicateResponseLongText(body) => {
                 let mut response = response.status(206);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -798,7 +787,7 @@ async fn multiple_auth_scheme_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = multiple_auth_scheme_get_validation();
 
@@ -818,7 +807,7 @@ where
 
     let resp = match result {
                                             Ok(rsp) => match rsp {
-                                                MultipleAuthSchemeGetResponse::Status200_CheckThatLimitingToMultipleRequiredAuthSchemesWorks
+                                                apis::default::MultipleAuthSchemeGetResponse::Status200_CheckThatLimitingToMultipleRequiredAuthSchemesWorks
                                                 => {
 
                                                   let mut response = response.status(200);
@@ -852,7 +841,7 @@ async fn one_of_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = one_of_get_validation();
 
@@ -869,7 +858,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            OneOfGetResponse::Status200_Success(body) => {
+            apis::default::OneOfGetResponse::Status200_Success(body) => {
                 let mut response = response.status(200);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -920,7 +909,7 @@ async fn override_server_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = override_server_get_validation();
 
@@ -940,7 +929,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            OverrideServerGetResponse::Status204_Success => {
+            apis::default::OverrideServerGetResponse::Status204_Success => {
                 let mut response = response.status(204);
                 response.body(Body::empty())
             }
@@ -977,7 +966,7 @@ async fn paramget_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = paramget_get_validation(query_params);
 
@@ -997,7 +986,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            ParamgetGetResponse::Status200_JSONRsp(body) => {
+            apis::default::ParamgetGetResponse::Status200_JSONRsp(body) => {
                 let mut response = response.status(200);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -1048,7 +1037,7 @@ async fn readonly_auth_scheme_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = readonly_auth_scheme_get_validation();
 
@@ -1068,7 +1057,7 @@ where
 
     let resp = match result {
                                             Ok(rsp) => match rsp {
-                                                ReadonlyAuthSchemeGetResponse::Status200_CheckThatLimitingToASingleRequiredAuthSchemeWorks
+                                                apis::default::ReadonlyAuthSchemeGetResponse::Status200_CheckThatLimitingToASingleRequiredAuthSchemeWorks
                                                 => {
 
                                                   let mut response = response.status(200);
@@ -1107,7 +1096,7 @@ async fn register_callback_post<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = register_callback_post_validation(query_params);
 
@@ -1127,7 +1116,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            RegisterCallbackPostResponse::Status204_OK => {
+            apis::default::RegisterCallbackPostResponse::Status204_OK => {
                 let mut response = response.status(204);
                 response.body(Body::empty())
             }
@@ -1168,7 +1157,7 @@ async fn required_octet_stream_put<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = required_octet_stream_put_validation(body);
 
@@ -1188,7 +1177,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            RequiredOctetStreamPutResponse::Status200_OK => {
+            apis::default::RequiredOctetStreamPutResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1220,7 +1209,7 @@ async fn responses_with_headers_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = responses_with_headers_get_validation();
 
@@ -1240,7 +1229,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            ResponsesWithHeadersGetResponse::Status200_Success {
+            apis::default::ResponsesWithHeadersGetResponse::Status200_Success {
                 body,
                 success_info,
                 bool_header,
@@ -1312,7 +1301,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            ResponsesWithHeadersGetResponse::Status412_PreconditionFailed {
+            apis::default::ResponsesWithHeadersGetResponse::Status412_PreconditionFailed {
                 further_info,
                 failure_info,
             } => {
@@ -1378,7 +1367,7 @@ async fn rfc7807_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = rfc7807_get_validation();
 
@@ -1395,7 +1384,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Rfc7807GetResponse::Status204_OK(body) => {
+            apis::default::Rfc7807GetResponse::Status204_OK(body) => {
                 let mut response = response.status(204);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -1418,7 +1407,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            Rfc7807GetResponse::Status404_NotFound(body) => {
+            apis::default::Rfc7807GetResponse::Status404_NotFound(body) => {
                 let mut response = response.status(404);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -1441,7 +1430,7 @@ where
                 .unwrap()?;
                 response.body(Body::from(body_content))
             }
-            Rfc7807GetResponse::Status406_NotAcceptable(body) => {
+            apis::default::Rfc7807GetResponse::Status406_NotAcceptable(body) => {
                 let mut response = response.status(406);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -1500,7 +1489,7 @@ async fn untyped_property_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = untyped_property_get_validation(body);
 
@@ -1519,18 +1508,20 @@ where
     let mut response = Response::builder();
 
     let resp = match result {
-        Ok(rsp) => match rsp {
-            UntypedPropertyGetResponse::Status200_CheckThatUntypedPropertiesWorks => {
-                let mut response = response.status(200);
-                response.body(Body::empty())
-            }
-        },
-        Err(_) => {
-            // Application code returned an error. This should not happen, as the implementation should
-            // return a valid response.
-            response.status(500).body(Body::empty())
-        }
-    };
+                                            Ok(rsp) => match rsp {
+                                                apis::default::UntypedPropertyGetResponse::Status200_CheckThatUntypedPropertiesWorks
+                                                => {
+
+                                                  let mut response = response.status(200);
+                                                  response.body(Body::empty())
+                                                },
+                                            },
+                                            Err(_) => {
+                                                // Application code returned an error. This should not happen, as the implementation should
+                                                // return a valid response.
+                                                response.status(500).body(Body::empty())
+                                            },
+                                        };
 
     resp.map_err(|e| {
         error!(error = ?e);
@@ -1552,7 +1543,7 @@ async fn uuid_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = uuid_get_validation();
 
@@ -1569,7 +1560,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            UuidGetResponse::Status200_DuplicateResponseLongText(body) => {
+            apis::default::UuidGetResponse::Status200_DuplicateResponseLongText(body) => {
                 let mut response = response.status(200);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -1627,7 +1618,7 @@ async fn xml_extra_post<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = xml_extra_post_validation(body);
 
@@ -1647,11 +1638,11 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            XmlExtraPostResponse::Status201_OK => {
+            apis::default::XmlExtraPostResponse::Status201_OK => {
                 let mut response = response.status(201);
                 response.body(Body::empty())
             }
-            XmlExtraPostResponse::Status400_BadRequest => {
+            apis::default::XmlExtraPostResponse::Status400_BadRequest => {
                 let mut response = response.status(400);
                 response.body(Body::empty())
             }
@@ -1690,7 +1681,7 @@ async fn xml_other_post<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = xml_other_post_validation(body);
 
@@ -1710,7 +1701,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            XmlOtherPostResponse::Status201_OK(body) => {
+            apis::default::XmlOtherPostResponse::Status201_OK(body) => {
                 let mut response = response.status(201);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
@@ -1726,7 +1717,7 @@ where
                 let body_content = body;
                 response.body(Body::from(body_content))
             }
-            XmlOtherPostResponse::Status400_BadRequest => {
+            apis::default::XmlOtherPostResponse::Status400_BadRequest => {
                 let mut response = response.status(400);
                 response.body(Body::empty())
             }
@@ -1765,7 +1756,7 @@ async fn xml_other_put<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = xml_other_put_validation(body);
 
@@ -1785,11 +1776,11 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            XmlOtherPutResponse::Status201_OK => {
+            apis::default::XmlOtherPutResponse::Status201_OK => {
                 let mut response = response.status(201);
                 response.body(Body::empty())
             }
-            XmlOtherPutResponse::Status400_BadRequest => {
+            apis::default::XmlOtherPutResponse::Status400_BadRequest => {
                 let mut response = response.status(400);
                 response.body(Body::empty())
             }
@@ -1828,7 +1819,7 @@ async fn xml_post<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = xml_post_validation(body);
 
@@ -1848,11 +1839,11 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            XmlPostResponse::Status201_OK => {
+            apis::default::XmlPostResponse::Status201_OK => {
                 let mut response = response.status(201);
                 response.body(Body::empty())
             }
-            XmlPostResponse::Status400_BadRequest => {
+            apis::default::XmlPostResponse::Status400_BadRequest => {
                 let mut response = response.status(400);
                 response.body(Body::empty())
             }
@@ -1891,7 +1882,7 @@ async fn xml_put<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     let validation = xml_put_validation(body);
 
@@ -1908,11 +1899,11 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            XmlPutResponse::Status201_OK => {
+            apis::default::XmlPutResponse::Status201_OK => {
                 let mut response = response.status(201);
                 response.body(Body::empty())
             }
-            XmlPutResponse::Status400_BadRequest => {
+            apis::default::XmlPutResponse::Status400_BadRequest => {
                 let mut response = response.status(400);
                 response.body(Body::empty())
             }
@@ -1957,7 +1948,7 @@ async fn create_repo<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::repo::Repo,
 {
     let validation = create_repo_validation(body);
 
@@ -1977,7 +1968,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            CreateRepoResponse::Status200_Success => {
+            apis::repo::CreateRepoResponse::Status200_Success => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -2014,7 +2005,7 @@ async fn get_repo_info<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::repo::Repo,
 {
     let validation = get_repo_info_validation(path_params);
 
@@ -2034,7 +2025,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            GetRepoInfoResponse::Status200_OK(body) => {
+            apis::repo::GetRepoInfoResponse::Status200_OK(body) => {
                 let mut response = response.status(200);
                 {
                     let mut response_headers = response.headers_mut().unwrap();
