+++ samples/server/petstore/rust-axum/output/ops-v3/src/server/mod.rs
@@ -10,24 +10,13 @@ use validator::{Validate, ValidationErrors};
 use crate::{header, types::*};
 
 #[allow(unused_imports)]
-use crate::models;
-
-use crate::{
-    Api, Op10GetResponse, Op11GetResponse, Op12GetResponse, Op13GetResponse, Op14GetResponse,
-    Op15GetResponse, Op16GetResponse, Op17GetResponse, Op18GetResponse, Op19GetResponse,
-    Op1GetResponse, Op20GetResponse, Op21GetResponse, Op22GetResponse, Op23GetResponse,
-    Op24GetResponse, Op25GetResponse, Op26GetResponse, Op27GetResponse, Op28GetResponse,
-    Op29GetResponse, Op2GetResponse, Op30GetResponse, Op31GetResponse, Op32GetResponse,
-    Op33GetResponse, Op34GetResponse, Op35GetResponse, Op36GetResponse, Op37GetResponse,
-    Op3GetResponse, Op4GetResponse, Op5GetResponse, Op6GetResponse, Op7GetResponse, Op8GetResponse,
-    Op9GetResponse,
-};
+use crate::{apis, models};
 
 /// Setup API Server.
 pub fn new<I, A>(api_impl: I) -> Router
 where
     I: AsRef<A> + Clone + Send + Sync + 'static,
-    A: Api + 'static,
+    A: apis::default::Default + 'static,
 {
     // build our application with a route
     Router::new()
@@ -85,7 +74,7 @@ async fn op10_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op10_get_validation())
@@ -105,7 +94,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op10GetResponse::Status200_OK => {
+            apis::default::Op10GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -137,7 +126,7 @@ async fn op11_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op11_get_validation())
@@ -157,7 +146,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op11GetResponse::Status200_OK => {
+            apis::default::Op11GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -189,7 +178,7 @@ async fn op12_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op12_get_validation())
@@ -209,7 +198,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op12GetResponse::Status200_OK => {
+            apis::default::Op12GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -241,7 +230,7 @@ async fn op13_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op13_get_validation())
@@ -261,7 +250,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op13GetResponse::Status200_OK => {
+            apis::default::Op13GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -293,7 +282,7 @@ async fn op14_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op14_get_validation())
@@ -313,7 +302,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op14GetResponse::Status200_OK => {
+            apis::default::Op14GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -345,7 +334,7 @@ async fn op15_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op15_get_validation())
@@ -365,7 +354,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op15GetResponse::Status200_OK => {
+            apis::default::Op15GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -397,7 +386,7 @@ async fn op16_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op16_get_validation())
@@ -417,7 +406,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op16GetResponse::Status200_OK => {
+            apis::default::Op16GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -449,7 +438,7 @@ async fn op17_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op17_get_validation())
@@ -469,7 +458,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op17GetResponse::Status200_OK => {
+            apis::default::Op17GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -501,7 +490,7 @@ async fn op18_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op18_get_validation())
@@ -521,7 +510,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op18GetResponse::Status200_OK => {
+            apis::default::Op18GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -553,7 +542,7 @@ async fn op19_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op19_get_validation())
@@ -573,7 +562,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op19GetResponse::Status200_OK => {
+            apis::default::Op19GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -605,7 +594,7 @@ async fn op1_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op1_get_validation())
@@ -625,7 +614,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op1GetResponse::Status200_OK => {
+            apis::default::Op1GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -657,7 +646,7 @@ async fn op20_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op20_get_validation())
@@ -677,7 +666,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op20GetResponse::Status200_OK => {
+            apis::default::Op20GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -709,7 +698,7 @@ async fn op21_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op21_get_validation())
@@ -729,7 +718,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op21GetResponse::Status200_OK => {
+            apis::default::Op21GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -761,7 +750,7 @@ async fn op22_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op22_get_validation())
@@ -781,7 +770,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op22GetResponse::Status200_OK => {
+            apis::default::Op22GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -813,7 +802,7 @@ async fn op23_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op23_get_validation())
@@ -833,7 +822,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op23GetResponse::Status200_OK => {
+            apis::default::Op23GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -865,7 +854,7 @@ async fn op24_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op24_get_validation())
@@ -885,7 +874,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op24GetResponse::Status200_OK => {
+            apis::default::Op24GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -917,7 +906,7 @@ async fn op25_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op25_get_validation())
@@ -937,7 +926,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op25GetResponse::Status200_OK => {
+            apis::default::Op25GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -969,7 +958,7 @@ async fn op26_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op26_get_validation())
@@ -989,7 +978,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op26GetResponse::Status200_OK => {
+            apis::default::Op26GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1021,7 +1010,7 @@ async fn op27_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op27_get_validation())
@@ -1041,7 +1030,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op27GetResponse::Status200_OK => {
+            apis::default::Op27GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1073,7 +1062,7 @@ async fn op28_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op28_get_validation())
@@ -1093,7 +1082,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op28GetResponse::Status200_OK => {
+            apis::default::Op28GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1125,7 +1114,7 @@ async fn op29_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op29_get_validation())
@@ -1145,7 +1134,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op29GetResponse::Status200_OK => {
+            apis::default::Op29GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1177,7 +1166,7 @@ async fn op2_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op2_get_validation())
@@ -1197,7 +1186,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op2GetResponse::Status200_OK => {
+            apis::default::Op2GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1229,7 +1218,7 @@ async fn op30_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op30_get_validation())
@@ -1249,7 +1238,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op30GetResponse::Status200_OK => {
+            apis::default::Op30GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1281,7 +1270,7 @@ async fn op31_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op31_get_validation())
@@ -1301,7 +1290,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op31GetResponse::Status200_OK => {
+            apis::default::Op31GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1333,7 +1322,7 @@ async fn op32_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op32_get_validation())
@@ -1353,7 +1342,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op32GetResponse::Status200_OK => {
+            apis::default::Op32GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1385,7 +1374,7 @@ async fn op33_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op33_get_validation())
@@ -1405,7 +1394,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op33GetResponse::Status200_OK => {
+            apis::default::Op33GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1437,7 +1426,7 @@ async fn op34_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op34_get_validation())
@@ -1457,7 +1446,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op34GetResponse::Status200_OK => {
+            apis::default::Op34GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1489,7 +1478,7 @@ async fn op35_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op35_get_validation())
@@ -1509,7 +1498,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op35GetResponse::Status200_OK => {
+            apis::default::Op35GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1541,7 +1530,7 @@ async fn op36_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op36_get_validation())
@@ -1561,7 +1550,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op36GetResponse::Status200_OK => {
+            apis::default::Op36GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1593,7 +1582,7 @@ async fn op37_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op37_get_validation())
@@ -1613,7 +1602,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op37GetResponse::Status200_OK => {
+            apis::default::Op37GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1645,7 +1634,7 @@ async fn op3_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op3_get_validation())
@@ -1665,7 +1654,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op3GetResponse::Status200_OK => {
+            apis::default::Op3GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1697,7 +1686,7 @@ async fn op4_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op4_get_validation())
@@ -1717,7 +1706,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op4GetResponse::Status200_OK => {
+            apis::default::Op4GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1749,7 +1738,7 @@ async fn op5_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op5_get_validation())
@@ -1769,7 +1758,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op5GetResponse::Status200_OK => {
+            apis::default::Op5GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1801,7 +1790,7 @@ async fn op6_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op6_get_validation())
@@ -1821,7 +1810,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op6GetResponse::Status200_OK => {
+            apis::default::Op6GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1853,7 +1842,7 @@ async fn op7_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op7_get_validation())
@@ -1873,7 +1862,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op7GetResponse::Status200_OK => {
+            apis::default::Op7GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1905,7 +1894,7 @@ async fn op8_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op8_get_validation())
@@ -1925,7 +1914,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op8GetResponse::Status200_OK => {
+            apis::default::Op8GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
@@ -1957,7 +1946,7 @@ async fn op9_get<I, A>(
 ) -> Result<Response, StatusCode>
 where
     I: AsRef<A> + Send + Sync,
-    A: Api,
+    A: apis::default::Default,
 {
     #[allow(clippy::redundant_closure)]
     let validation = tokio::task::spawn_blocking(move || op9_get_validation())
@@ -1977,7 +1966,7 @@ where
 
     let resp = match result {
         Ok(rsp) => match rsp {
-            Op9GetResponse::Status200_OK => {
+            apis::default::Op9GetResponse::Status200_OK => {
                 let mut response = response.status(200);
                 response.body(Body::empty())
             }
