+++ samples/server/petstore/rust-axum/output/petstore-with-fake-endpoints-models-for-testing/src/apis/fake_classname_tags123.rs
@@ -1,0 +1,32 @@
+use async_trait::async_trait;
+use axum::extract::*;
+use axum_extra::extract::{CookieJar, Multipart};
+use bytes::Bytes;
+use http::Method;
+use serde::{Deserialize, Serialize};
+
+use crate::{models, types::*};
+
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[must_use]
+#[allow(clippy::large_enum_variant)]
+pub enum TestClassnameResponse {
+    /// successful operation
+    Status200_SuccessfulOperation(models::Client),
+}
+
+/// FakeClassnameTags123
+#[async_trait]
+#[allow(clippy::ptr_arg)]
+pub trait FakeClassnameTags123 {
+    /// To test class name in snake case.
+    ///
+    /// TestClassname - PATCH /v2/fake_classname_test
+    async fn test_classname(
+        &self,
+        method: Method,
+        host: Host,
+        cookies: CookieJar,
+        body: models::Client,
+    ) -> Result<TestClassnameResponse, String>;
+}
