+++ samples/server/petstore/rust-axum/output/petstore-with-fake-endpoints-models-for-testing/src/apis/mod.rs
@@ -1,0 +1,6 @@
+pub mod another_fake;
+pub mod fake;
+pub mod fake_classname_tags123;
+pub mod pet;
+pub mod store;
+pub mod user;
