+++ samples/server/petstore/rust-axum/output/petstore/.openapi-generator/FILES
@@ -1,6 +1,10 @@
 .gitignore
 Cargo.toml
 README.md
+src/apis/mod.rs
+src/apis/pet.rs
+src/apis/store.rs
+src/apis/user.rs
 src/header.rs
 src/lib.rs
 src/models.rs
