+++ samples/server/petstore/rust-axum/output/petstore/src/apis/mod.rs
@@ -1,0 +1,3 @@
+pub mod pet;
+pub mod store;
+pub mod user;
