+++ samples/server/petstore/rust-axum/output/ping-bearer-auth/src/apis/mod.rs
@@ -1,0 +1,1 @@
+pub mod default;
