+++ samples/server/petstore/rust-axum/output/rust-axum-header-uuid/.openapi-generator/FILES
@@ -1,7 +1,8 @@
 .gitignore
-.openapi-generator-ignore
 Cargo.toml
 README.md
+src/apis/default.rs
+src/apis/mod.rs
 src/header.rs
 src/lib.rs
 src/models.rs
