+++ samples/server/petstore/rust-axum/output/rust-axum-header-uuid/src/apis/mod.rs
@@ -1,0 +1,1 @@
+pub mod default;
