+++ samples/server/petstore/rust-axum/output/rust-axum-test/src/apis/mod.rs
@@ -1,0 +1,1 @@
+pub mod default;
