+++ samples/server/petstore/rust-axum/output/rust-axum-validation-test/.openapi-generator/FILES
@@ -1,6 +1,8 @@
 .gitignore
 Cargo.toml
 README.md
+src/apis/default.rs
+src/apis/mod.rs
 src/header.rs
 src/lib.rs
 src/models.rs
