+++ samples/server/petstore/rust-axum/output/rust-axum-validation-test/src/apis/mod.rs
@@ -1,0 +1,1 @@
+pub mod default;
